Factorize request timeout setting into a new function request_set_timeout().
