+++ crates/nu-command/src/network/http/get.rs
@@ -1,5 +1,6 @@
 use crate::network::http::client::{
-    http_client, request_add_authorization_header, request_add_custom_headers, response_to_buffer,
+    http_client, request_add_authorization_header, request_add_custom_headers, request_set_timeout,
+    response_to_buffer,
 };
 use nu_engine::CallExt;
 use nu_protocol::ast::Call;
@@ -10,7 +11,6 @@ use nu_protocol::{
 use reqwest::StatusCode;
 use std::path::PathBuf;
 use std::str::FromStr;
-use std::time::Duration;
 
 #[derive(Clone)]
 pub struct SubCommand;
@@ -170,19 +170,7 @@ fn helper(
     let client = http_client(args.insecure.is_some());
     let mut request = client.get(url);
 
-    if let Some(timeout) = timeout {
-        let val = timeout.as_i64()?;
-        if val.is_negative() || val < 1 {
-            return Err(ShellError::TypeMismatch(
-                "Timeout value must be an integer and larger than 0".to_string(),
-                // timeout is already guaranteed to not be an error
-                timeout.expect_span(),
-            ));
-        }
-
-        request = request.timeout(Duration::from_secs(val as u64));
-    }
-
+    request = request_set_timeout(timeout, request)?;
     request = request_add_authorization_header(user, password, request);
     request = request_add_custom_headers(headers, request)?;
